disable flinging on preview view - only change camera mode when fling happens on the bottom tab - so zoom gestures do not interfere with the fling
